pyproject.toml
[tool.poetry]
name = "megaphone"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
youtube-dl = "^2021.1.24"
python-decouple = "^3.4"
flask = "^1.1.2"
validators = "^0.18.2"
pynacl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
